refactor(spelling): :recycle: replace Reqeust to Request
