+++ github/helpers/basicQueryMessage.ts
@@ -7,7 +7,7 @@ import {
 import { IRoom } from "@rocket.chat/apps-engine/definition/rooms";
 import { issueListMessage } from "../lib/issuesListMessage";
 import { contributorListMessage } from "../lib/contributorListMessage";
-import { pullRequestListMessage } from "../lib/pullReqeustListMessage";
+import { pullRequestListMessage } from "../lib/pullRequestListMessage";
 import { repoDataMessage } from "../lib/repoDataMessage";
 import { helperMessage } from "../lib/helperMessage";
 import { IAuthData } from "@rocket.chat/apps-engine/definition/oauth2/IOAuth2";
